+++ tests/testthat/test_weather.R
@@ -1,5 +1,6 @@
-library(envirocan)
-context("Weather data")
+
+# weather_dl --------------------------------------------------------------
+context("weather_dl")
 
 test_that("weather_dl (hour) downloads a data frame", {
   expect_silent(wd <- weather_dl(station_id = 51423, date = as.Date("2014-01-01"), skip = 15, interval = "hour"))
@@ -36,11 +37,11 @@ test_that("weather_dl (month) downloads a data frame", {
 })
 
 
-#####################
-## HOURLY
-#####################
 
-test_that("weather(hour) returns a data frame", {
+# weather by hour ---------------------------------------------------------
+context("weather by hour")
+
+test_that("weather (hour) returns a data frame", {
   expect_silent({weather(station_ids = 51423, start = "2014-01-01", end = "2014-01-31")})
   expect_message({w <- weather(station_ids = 51423, start = "2014-01-01", end = "2014-01-31", verbose = TRUE)})
 
@@ -64,13 +65,13 @@ test_that("weather(hour) returns a data frame", {
   expect_equal(w$qual[1], "Partner data that is not subject to review by the National Climate Archives")
 })
 
-test_that("weather(hourly) formats timezone display", {
+test_that("weather (hour) formats timezone display", {
   expect_silent({w <- weather(station_ids = 51423, start = "2014-03-01", end = "2014-04-01", tz_disp = "America/Vancouver")})
   expect_equal(w$time[1], as.POSIXct("2014-03-01 00:00:00", tz = "America/Vancouver"))
   expect_equal(w$time[w$date == as.Date("2014-04-01")][1], as.POSIXct("2014-04-01 01:00:00", tz = "America/Vancouver"))
 })
 
-test_that("weather(hourly) gets all", {
+test_that("weather (hour) gets all", {
   expect_silent(w <- weather(station_ids = 10816, start = Sys.Date() - lubridate::days(1), interval = "hour", trim = FALSE))
   expect_is(w, "data.frame")
   expect_length(w, 35)
@@ -79,11 +80,11 @@ test_that("weather(hourly) gets all", {
   expect_equal(w$date[nrow(w)], Sys.Date())
 })
 
-test_that("weather(hourly) trims NAs", {
+test_that("weather (hour) trims NAs", {
 
 })
 
-test_that("weather(hour) multiple stations", {
+test_that("weather (hour) multiple stations", {
   expect_silent({w <- weather(station_ids = c(51423, 10816), start = "2014-03-01", end = "2014-04-01")})
 
   expect_equal(unique(w$station_name), c("KAMLOOPS A", "SQUAMISH AIRPORT"))
@@ -91,9 +92,51 @@ test_that("weather(hour) multiple stations", {
 })
 
 
-#####################
-## DAILY
-#####################
+test_that("weather (hour) no data fails nicely", {
+  expect_silent(expect_message(w0 <- weather(1274, interval = "hour",
+                                             start = "2012-11-01", end = "2012-12-01"),
+                               "There are no data for station 1274 for interval by 'hour'."))
+  expect_silent(expect_message(w1 <- weather(c(1274, 1275), interval = "hour",
+                                             start = "2012-11-01", end = "2012-12-01"),
+                               "There are no data for station 1274 for interval by 'hour'."))
+
+  expect_is(w0, "data.frame")
+  expect_length(w0, 0)
+  expect_equal(nrow(w0), 0)
+  expect_is(w1, "data.frame")
+  expect_length(w1, 35)
+  expect_equal(nrow(w1), 744)
+
+  expect_silent(expect_message(w0 <- weather(1275, interval = "hour",
+                                             start = "2017-01-01", end = "2017-02-01"),
+                               "There are no data for these stations \\(1275\\) in this time range \\(2017-01-01 to 2017-02-01\\)."))
+
+  expect_silent(expect_message(w1 <- weather(c(1275, 1001), interval = "hour",
+                                             start = "2017-01-01", end = "2017-02-01"),
+                               "There are no data for these stations \\(1275, 1001\\) in this time range \\(2017-01-01 to 2017-02-01\\)."))
+
+  expect_is(w0, "data.frame")
+  expect_length(w0, 0)
+  expect_equal(nrow(w0), 0)
+  expect_is(w1, "data.frame")
+  expect_length(w1, 0)
+  expect_equal(nrow(w1), 0)
+})
+
+test_that("weather (hour) verbose and quiet", {
+  expect_message(weather(c(1275, 1001), interval = "hour",
+                         start = "2017-01-01", end = "2017-02-01"))
+
+  expect_silent(weather(c(1275, 1001), interval = "hour",
+                         start = "2017-01-01", end = "2017-02-01", quiet = TRUE))
+
+  expect_message(weather(c(1275, 1001), interval = "hour",
+                        start = "2017-01-01", end = "2017-02-01", verbose = TRUE),
+                 "(Getting station: 1275\\n)")
+})
+
+# weather by day ----------------------------------------------------------
+context("weather by day")
 
 test_that("weather (day) returns a data frame", {
   expect_message(expect_error({weather(station_ids = 51423, start = "2014-01-01", end = "2014-03-01", interval = "day")}, NA))
@@ -117,7 +160,7 @@ test_that("weather (day) returns a data frame", {
   expect_equal(w$qual[1], "Partner data that is not subject to review by the National Climate Archives")
 })
 
-test_that("weather(daily) gets all", {
+test_that("weather (day) gets all", {
   expect_silent(weather(station_ids = 54398, interval = "day", trim = FALSE))
   expect_message(expect_error({w <- weather(station_ids = 54398, interval = "day", trim = FALSE, verbose = TRUE)}, NA))
 
@@ -138,9 +181,7 @@ test_that("weather(daily) gets all", {
 })
 
 
-
-
-test_that("weather(daily) trims NAs", {
+test_that("weather (day) trims NAs", {
   expect_silent(w1 <- weather(station_ids = 54398, interval = "day", trim = FALSE))
   expect_silent(w2 <- weather(station_ids = 54398, interval = "day", trim = TRUE))
 
@@ -148,28 +189,58 @@ test_that("weather(daily) trims NAs", {
   expect_gte(length(data.frame(w1)[is.na(data.frame(w1))]), length(data.frame(w2)[is.na(data.frame(w2))]))
 })
 
-test_that("weather(day) no data fails nicely", {
-  expect_message(expect_error(w0 <- weather(station_ids = 51457, start = "2014-01-01", end = "2014-05-01", interval = "day"), NA))
-
-  ## Basics
-  expect_is(w0, "data.frame")
-  expect_length(w0, 0)
-  expect_equal(nrow(w0), 0)
-})
-
-test_that("weather(day) mutliple stations", {
+test_that("weather (day) mutliple stations", {
   expect_error(w <- weather(station_ids = c(54398, 51423), start = "2016-03-01", end = "2016-04-01", interval = "day"), NA)
 
   expect_equal(unique(w$station_name), c("MUSKOKA SNOW", "KAMLOOPS A"))
   expect_equal(nrow(w[w$station_id == 54398,]), nrow(w[w$station_id == 51423,]))
 })
 
+test_that("weather (day) no data fails nicely", {
+  expect_silent(expect_message(w0 <- weather(station_ids = 42013, interval = "day",
+                                             start = "2017-01-01", end = "2017-02-01"),
+                               "There are no data for station 42013 for interval by 'day'."))
+  expect_silent(expect_message(w1 <- weather(station_ids = c(42013, 51423), interval = "day",
+                                             start = "2017-01-01", end = "2017-02-01"),
+                               "There are no data for station 42013 for interval by 'day'."))
 
-#####################
-## MONTHLY
-#####################
+  expect_is(w0, "data.frame")
+  expect_length(w0, 0)
+  expect_equal(nrow(w0), 0)
+  expect_is(w1, "data.frame")
+  expect_length(w1, 36)
+  expect_equal(nrow(w1), 32)
 
-test_that("weather returns a data frame MONTHLY", {
+  expect_silent(expect_message(w0 <- weather(1274, interval = "day", start = "2017-01-01", end = "2017-03-01"),
+                               "There are no data for these stations \\(1274\\) in this time range \\(2017-01-01 to 2017-03-01\\)."))
+  expect_silent(expect_message(w1 <- weather(c(1274, 1275), interval = "day", start = "2017-01-01", end = "2017-03-01"),
+                               "There are no data for these stations \\(1274, 1275\\) in this time range \\(2017-01-01 to 2017-03-01\\)."))
+
+  ## Basics
+  expect_is(w0, "data.frame")
+  expect_length(w0, 0)
+  expect_equal(nrow(w0), 0)
+  expect_is(w1, "data.frame")
+  expect_length(w1, 0)
+  expect_equal(nrow(w1), 0)
+})
+
+test_that("weather (day) verbose and quiet", {
+  expect_message(weather(c(42013, 51423), interval = "day",
+                         start = "2017-01-01", end = "2017-02-01"))
+
+  expect_silent(weather(c(42013, 51423), interval = "day",
+                        start = "2017-01-01", end = "2017-02-01", quiet = TRUE))
+
+  expect_message(weather(c(42013, 51423), interval = "day",
+                         start = "2017-01-01", end = "2017-02-01", verbose = TRUE),
+                 "(Getting station: 42013\\n)")
+})
+
+# weather by month --------------------------------------------------------
+context("weather by month")
+
+test_that("weather (month) returns a data frame", {
   expect_silent(w <- weather(station_ids = 5401, start = "2014-01-01", end = "2014-05-01", interval = "month"))
 
   ## Basics
@@ -193,35 +264,64 @@ test_that("weather returns a data frame MONTHLY", {
 
 })
 
-test_that("weather(monthly) trims NAs", {
+test_that("weather (month) trims NAs", {
 
 })
 
-test_that("weather(monthly) no data fails nicely", {
-  expect_message(expect_error(w0 <- weather(station_ids = 51423, start = "2014-01-01", end = "2014-05-01", interval = "month"), NA))
-
-  ## Basics
+test_that("weather (month) no data fails nicely", {
+  expect_silent(expect_message(w0 <- weather(station_ids = 51423, interval = "month",
+                                             start = "2012-01-01", end = "2012-02-01"),
+                               "There are no data for station 51423 for interval by 'month'."))
+  expect_silent(expect_message(w1 <- weather(station_ids = c(51423, 1275), interval = "month",
+                                             start = "2012-01-01", end = "2012-02-01"),
+                               "There are no data for station 51423 for interval by 'month'."))
   expect_is(w0, "data.frame")
   expect_length(w0, 0)
   expect_equal(nrow(w0), 0)
+  expect_is(w1, "data.frame")
+  expect_length(w1, 34)
+  expect_equal(nrow(w1), 2)
+
+  expect_silent(expect_message(w0 <- weather(1274, interval = "month", start = "2017-01-01", end = "2017-02-01"),
+                               "There are no data for these stations \\(1274\\) in this time range \\(2017-01-01 to 2017-02-01\\)."))
+  expect_silent(expect_message(w1 <- weather(c(1274, 1275), interval = "month", start = "2017-01-01", end = "2017-02-01"),
+                               "There are no data for these stations \\(1274, 1275\\) in this time range \\(2017-01-01 to 2017-02-01\\)."))
+
+  expect_is(w0, "data.frame")
+  expect_length(w0, 0)
+  expect_equal(nrow(w0), 0)
+  expect_is(w1, "data.frame")
+  expect_length(w1, 0)
+  expect_equal(nrow(w1), 0)
 })
 
-test_that("weather(month) multiple stations", {
+test_that("weather (month) multiple stations", {
   expect_silent(w <- weather(station_ids = c(5401, 5940), start = "2014-01-01", end = "2014-05-01", interval = "month"))
 
   expect_equal(unique(w$station_name), c("MAGOG", "ST PRIME"))
   expect_equal(nrow(w[w$station_id == 5401,]), nrow(w[w$station_id == 5940,]))
 })
 
-test_that("weather(month) multiple stations (one NA)", {
+test_that("weather (month) multiple stations (one NA)", {
   expect_message(expect_error(w <- weather(station_ids = c(5401, 51423), start = "2014-01-01", end = "2014-05-01", interval = "month"), NA))
   expect_equal(unique(w$station_name), c("MAGOG"))
 })
 
+test_that("weather (month) verbose and quiet", {
+  expect_message(weather(c(51423, 1275), interval = "month",
+                         start = "2017-01-01", end = "2017-02-01"))
 
-##################
-### list_col
-##################
+  expect_silent(weather(c(51423, 1275), interval = "month",
+                        start = "2017-01-01", end = "2017-02-01", quiet = TRUE))
+
+  expect_message(weather(c(51423, 1275), interval = "month",
+                         start = "2017-01-01", end = "2017-02-01", verbose = TRUE),
+                 "(Getting station: 51423\\n)")
+})
+
+
+# list_cols ---------------------------------------------------------------
+
 context("Generating list_col")
 
 test_that("list_col=TRUE and interval=hour groups on the right level", {