+++ NEWS.md
@@ -3,6 +3,25 @@ title: "News"
 author: "Steffi LaZerte"
 output: html_document
 ---
+
+# weathercan 0.2.2
+
+## Changes
+- Update and expand vignettes (closes #15)
+- Data now returned as tibbles
+- Added listcol functionality (closes #8)
+- Added internal tests for interpolation
+- Updated R version
+- Standardized reference to stations dataset (`stn`) in all functions
+
+## Major changes
+- envirocan renamed to weathercan (closes #17)
+
+## Bug fixes
+- Fixed inclusion of New Brunswick stations (closes #9)
+- Downloads with no data return empty tibble and an informative message (closes #21)
+
+
 # envirocan 0.2.1 (2017-03-04)
 - Minor bug fixes: correcting encoding information for downloads, updating function calls to dplyr package, updating stations dataset